pyproject.toml
[tool.poetry]
name = "drf_clean_project"
version = "0.1.0"
description = ""
authors = ["noskriptt <maryaskin.a.v@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
coreapi = "^2.3.3"
django-cors-headers = "^3.13.0"
django-filter = "^22.1"
Django = "3.2"
djangorestframework-simplejwt = "^5.2.0"
djangorestframework = "^3.13.1"
drf-yasg = "^1.21.3"
environs = "^9.5.0"
gunicorn = "^20.1.0"
Markdown = "^3.4.1"
Pillow = "^9.2.0"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
